+++ app/renderer/pages/homePage/index.js
@@ -3,7 +3,6 @@ import Button from 'components/Button';
 import styled from 'styled-components';
 import Spinner from 'components/Spinner';
 import { Layout, LayoutWrapper, MainContent } from 'components/layout';
-import { cennznetStatusChannel } from 'renderer/ipc/cennznet.ipc';
 import SideNav from 'components/layout/SideNav';
 import SimpleSidebar from 'components/layout/SimpleSidebar';
 import { Logger } from 'renderer/utils/logging';